Fixed warnings, many require logic revision though.

+++ src/gfx/font/font.c
@@ -10,7 +10,7 @@ void draw_char(ca_layer* layer, char ch, int x, int y, Color color) {
 	Coordinate p = point_make(x, y);
 	if (p.x < 0 || p.y < 0 || p.x >= layer->size.width || p.y >= layer->size.height) return;
 
-	int* bitmap = font8x8_basic[ch];
+	int* bitmap = font8x8_basic[(int)ch];
 	for (int i = 0; i < CH_H; i++) {
 		char row = bitmap[i];
 		for (int j = 0; j < CH_W; j++) {
@@ -20,5 +20,3 @@ void draw_char(ca_layer* layer, char ch, int x, int y, Color color) {
 		}
 	}
 }
-
-
